requirements.txt
matplotlib==3.7.1
nltk==3.8.1
numpy==1.24.3
pandas==1.5.3
Pillow==10.1.0
seaborn==0.13.0
streamlit==1.28.1
textblob==0.17.1
wordcloud==1.9.2


